requirements.txt
fastapi
uvicorn
openai
twilio
python-dotenv
python-multipart
redis
pydub
redisvl
numpy
aiofiles
requests